done the stock

+++ app/src/main/java/com/example/watchstore/CartItem.kt
@@ -2,7 +2,7 @@ package com.example.watchstore
 
 data class CartItem(
     val productId: String = "",
-    val quantity: Int = 0,
+    val quantity: Long = 0,
     val imageUrl: String = "",
-    val price: Int = 0
+    val price: Double = 0.0
 )
